[CHORE] #104 : analyzeText, analyzeImage -> textInput, imageInput rename
